requirements.txt
sentence-transformers
scikit-learn
transformers
torch
openai
python-docx
sentence-transformers
numpy
pandas
faiss-cpu
